requirements.txt
requests>=2.31.0
requests-html>=0.10.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
selenium>=4.15.0
lxml>=4.9.0
webdriver-manager>=4.0.0
colorlog>=6.7.0
streamlit>=1.39.0
